refactor(hooks): move hooks.ts into lib/hooks/install.ts
Third hub-drain slice: the registrar already belongs next to cache.ts,
match.ts, and profile.ts. No behavior change — import retargets and
session-tracker path walks gain one hop for dist/lib/hooks/. No hub
re-export shim.
