+++ apps/cli/src/lib/hooks/install.test.ts
@@ -49,7 +49,7 @@ interface VersionHooksReport {
 }
 
 function runVersionHooksInventory(agent: string, version: string): VersionHooksReport {
-  const modulePath = path.resolve(process.cwd(), 'src/lib/hooks.ts');
+  const modulePath = path.resolve(process.cwd(), 'src/lib/hooks/install.ts');
   const script = `
     const mod = await import(${JSON.stringify(modulePath)});
     console.log(JSON.stringify({
@@ -154,7 +154,7 @@ function seedClaudeVersionWithGeneratedShim(version: string, hookName: string, e
 /** Run checkVersionHookWiring (optionally after registerHooksToSettings) in a
  *  subprocess rooted at testHome, returning its JSON report. */
 function runWiring(agent: string, version: string, opts: { register?: boolean } = {}): WiringReport {
-  const modulePath = path.resolve(process.cwd(), 'src/lib/hooks.ts');
+  const modulePath = path.resolve(process.cwd(), 'src/lib/hooks/install.ts');
   const versionsPath = path.resolve(process.cwd(), 'src/lib/installations/versions.ts');
   const register = opts.register
     ? `
@@ -279,7 +279,7 @@ interface RuntimeRepairReport {
 }
 
 function runRuntime(scriptBody: string): string {
-  const modulePath = path.resolve(process.cwd(), 'src/lib/hooks.ts');
+  const modulePath = path.resolve(process.cwd(), 'src/lib/hooks/install.ts');
   const script = `
     const mod = await import(${JSON.stringify(modulePath)});
     ${scriptBody}
@@ -631,7 +631,7 @@ describe('repairManagedHookRuntimeArtifacts', () => {
 
 describe('installSessionTrackerHookSync — failure reasons are never empty', () => {
   it('returns installed:false with a NON-EMPTY error when the hook cannot be installed', async () => {
-    const { installSessionTrackerHookSync } = await import('./hooks.js');
+    const { installSessionTrackerHookSync } = await import('./install.js');
     // antigravity passes the CLI-side supports(agent,'hooks') gate, but the
     // session-tracker child declines it (no SessionStart hook event) and prints
     // the reason to STDOUT. Pre-fix, err.stderr was an empty-but-truthy Buffer,
